services/ad-service/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
PyMySQL==1.1.0
redis==4.6.0
python-dotenv==1.0.0
Werkzeug==2.3.6
gunicorn==21.2.0
requests==2.31.0
Pillow==10.0.0
moviepy==1.0.3
python-magic==0.4.27